Standalone-components-was-set
