+++ src/app/app.component.html
@@ -6,10 +6,10 @@
     </div>
     <div class="row d-flex flex-column mt-5 pt-3">
         <div class="col col-lg-10 mx-auto">
-            <router-outlet></router-outlet>
+            <!-- <router-outlet></router-outlet> -->
         </div>
         <div class='col mx-auto px-0' style='position: relative; margin-bottom: 0;'>
-            <app-footer></app-footer>
+            <!-- <app-footer></app-footer> -->
         </div>
     </div>
 </main>
