+++ src/app/app.module.ts
@@ -1,55 +1,55 @@
 
-import { AppComponent } from './app.component';
-import { AppRoutingModule } from './app-routing.module';
-import { BrowserModule } from '@angular/platform-browser';
-import { environment } from './../environment/environment';
-import { FooterComponent } from './footer/footer.component';
-import { HeaderComponent } from './header/header.component';
+
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { NgModule } from '@angular/core';
 
-import { provideFirebaseApp, getApp, initializeApp } from '@angular/fire/app';
-import { getFirestore, provideFirestore } from '@angular/fire/firestore';
-import { provideAuth,getAuth } from '@angular/fire/auth';
-import { provideDatabase,getDatabase } from '@angular/fire/database';
-import { provideStorage,getStorage } from '@angular/fire/storage';
+// import { provideFirebaseApp, getApp, initializeApp } from '@angular/fire/app';
+// import { getFirestore, provideFirestore } from '@angular/fire/firestore';
+
+// import { provideStorage,getStorage } from '@angular/fire/storage';
 import { AngularFireModule } from '@angular/fire/compat';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { provideAuth,getAuth } from '@angular/fire/auth';
+import { provideDatabase,getDatabase } from '@angular/fire/database';
+import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+// import { BrowserModule } from '@angular/platform-browser';
+import { environment } from './../environment/environment';
+import { FooterComponent } from './footer/footer.component';
+import { HeaderComponent } from './header/header.component';
+// import { MatSliderModule } from '@angular/material/slider';
+// import { MatButtonModule } from '@angular/material/button';
+// import { MatInputModule } from '@angular/material/input';
 
-import { MatSliderModule } from '@angular/material/slider';
-import { MatButtonModule } from '@angular/material/button';
-import { MatInputModule } from '@angular/material/input';
+// const materialComponents = [
+//   // MatSliderModule,
+//   // MatButtonModule,
+//   // MatInputModule
+// ];
 
-const materialComponents = [
-  MatSliderModule,
-  MatButtonModule,
-  MatInputModule
-];
+// @NgModule({
+//   declarations: [
+//     HeaderComponent,
+//     FooterComponent
+//   ],
+//   imports: [
+//     // AngularFireAuthModule,
+//     // AngularFireStorageModule,
+//     // provideFirebaseApp.initializeApp(environment.firebaseConfig),
+//     AppRoutingModule,
+//     BrowserModule,
+//     // materialComponents,
+//     NgbModule,
+//     provideAuth(() => getAuth()),
+//     provideDatabase(() => getDatabase()),
+//     AngularFireModule.initializeApp(environment.firebaseConfig),
+//     BrowserAnimationsModule,
+//     // provideFirebaseApp(() => initializeApp(environment.firebaseConfig)),
+//     // provideFirestore(() => getFirestore()),
+//     // provideStorage(() => getStorage()),
+//   ],
+//   providers: [],
+//   // bootstrap: [AppComponent]
+// })
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    FooterComponent
-  ],
-  imports: [
-    // AngularFireAuthModule,
-    // AngularFireStorageModule,
-    // provideFirebaseApp.initializeApp(environment.firebaseConfig),
-    AppRoutingModule,
-    BrowserModule,
-    materialComponents,
-    NgbModule,
-    provideAuth(() => getAuth()),
-    provideDatabase(() => getDatabase()),
-    AngularFireModule.initializeApp(environment.firebaseConfig),
-    BrowserAnimationsModule,
-    // provideFirebaseApp(() => initializeApp(environment.firebaseConfig)),
-    // provideFirestore(() => getFirestore()),
-    // provideStorage(() => getStorage()),
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
-
-export class AppModule { }
+// export class AppModule { }
